tests/requirements-test.txt
# Test-only dependencies (separate from each plugin's runtime requirements)
pytest>=8.0
requests>=2.32
Pillow>=10.0.0
stashapi>=0.1.5
